feat: google oidc flow

+++ libs/database/tokens.go
@@ -64,7 +64,7 @@ func (m TokenModel) Insert(token *Token) error {
 	return err
 }
 
-func (m TokenModel) DeleteAllForUser(scope string, userID int64) error {
+func (m TokenModel) DeleteAllForUser(scope string, userID int) error {
 	query := `
         DELETE FROM tokens 
         WHERE scope = $1 AND user_id = $2`
